Product Controller moved back to package com.kodilla.ecommercee to resolve conflicts
