+++ src/main/java/com/kodilla/ecommercee/GenericEntity.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.domain;
+package com.kodilla.ecommercee;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
