+++ src/main/java/com/kodilla/ecommercee/repository/GenericEntityRepository.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee.repository;
 
-import com.kodilla.ecommercee.domain.GenericEntity;
+import com.kodilla.ecommercee.GenericEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
